fix(interface/progress): progressProps being set by client crashing users in range
By adding extra validation on clientside and using a hard limit for prop count associated with a progress bar (set via convar) avoids cheaters setting hundreds of props creating either massive performance issues or simply crashing all players.
This fixes it by setting the statebag serverside (allowing it to work with strict statebags) but also adds in an extra limit to the amount of props that can be set.

Co-authored-by: Linden <65407488+thelindat@users.noreply.github.com>